+++ src/init.js
@@ -4,7 +4,9 @@
     angular.module('djangularRestFramework', [
         'drf-config',
         'drf-provider',
-        'drf-display-name-filter'
+        'drf-display-name-filter',
+        'drf-field-errors',
+        'drf-form-field'
     ]);
 
 }(window.angular));